VCV-30: Move sync button to dashboard summary page
