+++ src/features/review/components/dashboard/page-client.tsx
@@ -1,13 +1,23 @@
 'use client';
 
-import { useMemo } from 'react';
+import { useMemo, useState } from 'react';
 import { Card, CardContent } from '@/ui/card';
+import { Button } from '@/ui/button';
 import { useDashboardDataQuery } from '@/features/review/hooks';
+import { useDhis2SyncMutation } from '@/features/review/hooks/use-dhis2-sync';
 import { getMonthDateRange } from '@/features/review/utils/master-table-view';
 import { SiteInformationSection } from './site-information-section';
 import { EntomologicalSummarySection } from './entomological-summary-section';
 import { BednetsDataSection } from './bednets-data-section';
 import { DashboardLoadingSkeleton } from './loading-skeleton';
+import {
+  useUserPermissionsQuery,
+  canPushSites,
+} from '@/features/user';
+import { showSuccessToast } from '@/shared/ui/show-success-toast';
+import { showErrorToast } from '@/shared/ui/show-error-toast';
+import { Dhis2SyncDialog } from '@/features/review/components/review-dashboard/dhis2-sync-dialog';
+import type { VillageIrsFormData, SiteIrsData } from '@/features/review/types/dhis2-sync';
 
 interface DashboardPageClientProps {
   district: string;
@@ -22,7 +32,13 @@ export function DashboardPageClient({
   const startDate = dateRange?.startDate;
   const endDate = dateRange?.endDate;
 
-  const [year, monthNum] = monthYear.split('-').map(Number);
+  const [yearString, monthString] = monthYear.split('-');
+  const year = Number.parseInt(yearString, 10);
+  const monthNum = Number.parseInt(monthString, 10);
+  const [dialogOpen, setDialogOpen] = useState(false);
+
+  const { data: permissions } = useUserPermissionsQuery();
+  const canSync = permissions ? canPushSites(permissions) : false;
 
   const { data, isLoading, error } = useDashboardDataQuery({
     district,
@@ -36,6 +52,114 @@ export function DashboardPageClient({
       year: 'numeric',
     });
   }, [year, monthNum]);
+
+  const uniqueVillages = useMemo(() => {
+    if (!permissions?.sites?.canAccessSites) return [];
+
+    const villages = new Set<string>();
+    permissions.sites.canAccessSites.forEach(site => {
+      if (site.villageName && site.district === district) {
+        villages.add(site.villageName);
+      }
+    });
+
+    return Array.from(villages).sort();
+  }, [permissions?.sites?.canAccessSites, district]);
+
+  const extractSummary = (payload: unknown) => {
+    if (!payload || typeof payload !== 'object') return null;
+    const summary = (payload as Record<string, unknown>).summary;
+    if (!summary || typeof summary !== 'object') return null;
+    const { totalHouseholds, successfulSyncs, failedSyncs, skippedHouseholds } =
+      summary as Record<string, unknown>;
+    if (
+      [totalHouseholds, successfulSyncs, failedSyncs, skippedHouseholds].every(
+        value => typeof value === 'number',
+      )
+    ) {
+      return {
+        totalHouseholds: totalHouseholds as number,
+        successfulSyncs: successfulSyncs as number,
+        failedSyncs: failedSyncs as number,
+        skippedHouseholds: skippedHouseholds as number,
+      };
+    }
+    return null;
+  };
+
+  const { mutateAsync: triggerSync, isPending: isSyncing } = useDhis2SyncMutation({
+    onSuccess: data => {
+      const summary = extractSummary(data);
+      if (summary) {
+        showSuccessToast(
+          `DHIS2 sync processed ${summary.totalHouseholds} household${
+            summary.totalHouseholds === 1 ? '' : 's'
+          }: ${summary.successfulSyncs} succeeded, ${summary.failedSyncs} failed, ${summary.skippedHouseholds} skipped.`,
+        );
+        return;
+      }
+      showSuccessToast(
+        'DHIS2 sync started. You will receive a notification when it finishes.',
+      );
+    },
+    onError: error => {
+      console.error('DHIS2 sync error:', error);
+      const errorMessage = error instanceof Error ? error.message : 'Failed to sync with DHIS2';
+      showErrorToast(errorMessage);
+    },
+  });
+
+  const handleSyncSubmit = (villageData: VillageIrsFormData[]) => {
+    if (!permissions?.sites?.canAccessSites) return;
+
+    const villageToIrsMap = new Map(
+      villageData.map(v => [v.villageName, v])
+    );
+
+    const irsData: SiteIrsData[] = permissions.sites.canAccessSites
+      .filter(site =>
+        site.villageName &&
+        site.district === district &&
+        villageToIrsMap.has(site.villageName)
+      )
+      .map(site => {
+        const villageIrs = villageToIrsMap.get(site.villageName!)!;
+        if (!villageIrs.wasIrsSprayed) {
+          return {
+            siteId: site.siteId,
+            wasIrsSprayed: false,
+          } satisfies SiteIrsData;
+        }
+        return {
+          siteId: site.siteId,
+          wasIrsSprayed: true,
+          insecticideSprayed: villageIrs.insecticideSprayed,
+          dateLastSprayed: villageIrs.dateLastSprayed,
+        } satisfies SiteIrsData;
+      });
+
+    console.log('Submitting DHIS2 sync with IRS data:', {
+      district,
+      year,
+      month: monthNum,
+      irsDataCount: irsData.length,
+      irsData,
+    });
+
+    setDialogOpen(false);
+
+    triggerSync({
+      district,
+      year,
+      month: monthNum,
+      dryRun: false,
+      irsData,
+    });
+  };
+
+  const handleSyncButtonClick = () => {
+    setDialogOpen(true);
+  };
 
   if (isLoading) {
     return <DashboardLoadingSkeleton />;
@@ -82,9 +206,25 @@ export function DashboardPageClient({
   return (
     <div className="min-h-screen bg-white">
       <div className="container mx-auto px-6 py-8">
-        <div className="mb-8">
-          <h1 className="text-4xl font-bold text-gray-800">{district}</h1>
-          <p className="mt-2 text-lg text-gray-600">{monthName}</p>
+        <div className="mb-8 flex flex-col gap-4 md:flex-row md:items-center md:justify-between">
+          <div>
+            <h1 className="text-4xl font-bold text-gray-800">{district}</h1>
+            <p className="mt-2 text-lg text-gray-600">{monthName}</p>
+          </div>
+          <div className="flex flex-col items-start gap-2 md:items-end">
+            <Button
+              className="bg-chart-green-medium text-white shadow-xs hover:bg-chart-green-dark focus-visible:border-chart-green-medium focus-visible:ring-chart-green-medium/40"
+              onClick={handleSyncButtonClick}
+              disabled={isSyncing || !canSync}
+            >
+              {isSyncing ? 'Syncing with DHIS2…' : 'Sync to DHIS2'}
+            </Button>
+            {!canSync && (
+              <p className="text-muted-foreground text-xs md:text-right">
+                You do not have permission to push metadata.
+              </p>
+            )}
+          </div>
         </div>
 
         <div className="space-y-12">
@@ -98,6 +238,14 @@ export function DashboardPageClient({
           <BednetsDataSection data={data.entomologicalSummary} />
         </div>
       </div>
+
+      <Dhis2SyncDialog
+        open={dialogOpen}
+        onOpenChange={setDialogOpen}
+        villages={uniqueVillages}
+        onSubmit={handleSyncSubmit}
+        isSubmitting={isSyncing}
+      />
     </div>
   );
 }
